Update README.md
Description:
Welcome to our global tours booking website using single page technology with react JS , JS ,HTML ,CSS and API.
your passport to a world of unforgettable adventures and experiences! We are your one-stop platform for planning and booking tours worldwide, designed to make your travel dreams come true.

Start your journey of a lifetime today! Explore our website, discover incredible tours, and embark on the adventure you've always dreamed of. The world is waiting for you, and we're here to help you make the most of it. Book with us and create memories that will last a lifetime.